Priceband table seeder and view done

+++ database/migrations/2020_08_10_111935_create_price_bands_table.php
@@ -18,7 +18,7 @@ class CreatePriceBandsTable extends Migration
             $table->integer('from')->unigned();
             $table->integer('to')->unsigned();
             $table->string('band_type');
-            $table->integer('cost')->unsigned();
+            $table->decimal('cost',8,2)->unsigned();
             $table->timestamps();
         });
     }
